Change percentage types to float

+++ lib/laa_crime_schemas/structs/property.rb
@@ -11,8 +11,8 @@ module LaaCrimeSchemas
       attribute :bedrooms, Types::Integer.optional
       attribute :value, Types::PenceSterling
       attribute :outstanding_mortgage, Types::PenceSterling
-      attribute :percentage_applicant_owned, Types::Integer
-      attribute :percentage_partner_owned, Types::Integer.optional
+      attribute :percentage_applicant_owned, Types::Float
+      attribute :percentage_partner_owned, Types::Float.optional
       attribute :is_home_address, Types::YesNoValue.optional
       attribute :has_other_owners, Types::YesNoValue
       attribute :address, Address.optional
